% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyHelpers.R
\name{BinomialHelper}
\alias{BinomialHelper}
\title{BinomialHelper}
\usage{
BinomialHelper(size, prob, SingletonList)
}
\arguments{
\item{size}{number of trials}

\item{prob}{probability of success on each trial}

\item{SingletonList}{Boolean indicating whether to return a single function as a list. Default FALSE.}
}
\value{
A list of functions.
}
\description{
This will return a list of poisson random number generator functions.
}
\details{
The lambda parameter may be a vector.
}
\examples{

myFuncs <- BinomialHelper(size = 100, prob = c(.9, .1))
myFuncs[[1]](4)

}
