% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateClaimsByEval.R
\name{ClaimsByEvaluationDate}
\alias{ClaimsByEvaluationDate}
\title{Claims by evaluation date}
\usage{
ClaimsByEvaluationDate(dfPolicy, Frequency, Severity, Link, EvaluationDates)
}
\arguments{
\item{dfPolicy}{A policy data frame}

\item{Frequency}{A function which will randomly generate number of claims}

\item{Severity}{A function which will randomly generate the severity of each claim}

\item{Link}{A function which dictates how severities change from one evaluation date to the next}

\item{EvaluationDates}{A vector of evaluation dates}
}
\value{
A list with revised policy table and a new claims table
}
\description{
Given a data frame of policies, this will simulate claims
}
\details{
Each function
}

