% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateClaimsByLag.R
\name{ClaimsByLag}
\alias{ClaimsByLag}
\title{Claims by lag}
\usage{
ClaimsByLag(dfPolicy, Frequency, Severity, Links, Lags)
}
\arguments{
\item{dfPolicy}{A policy data frame}

\item{Frequency}{A function which will randomly generate number of claims}

\item{Severity}{A function which will randomly generate the severity of each claim}

\item{Links}{A list of functions which dictates how severities change from one evaluation date to the next}

\item{Lags}{A vector of lags}
}
\value{
A claims data frame
}
\description{
Given a data frame of policies, this will simulate claims
}
\details{
The claim amounts are presumed to be cumulative
}
