% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaimsByWaitTime.R
\name{ClaimsByWaitTime}
\alias{ClaimsByWaitTime}
\title{ClaimsByWaitTime}
\usage{
ClaimsByWaitTime(dfPolicy, ClaimFrequency, PaymentFrequency, OccurrenceWait,
  ReportWait, PayWait, PaySeverity, PayOnlyPositive = TRUE)
}
\arguments{
\item{dfPolicy}{A data frame of policy records}

\item{ClaimFrequency}{A function which will randomly generate the number of claims per policy}

\item{PaymentFrequency}{A function which will determine the number of payments per claim}

\item{OccurrenceWait}{A function which will generate the time until occurrence for each claim}

\item{ReportWait}{A function which will generate the time until report}

\item{PayWait}{A function which will generate the lag time between payments}

\item{PaySeverity}{A function which will randomly generate the severity of each claim payment}

\item{PayOnlyPositive}{Boolean indicating whether to discard negative payments.}
}
\description{
Construct a data frame of claims simulated by time between events.
}
\details{
This function will generate claim transactions.

Wait times and frequencies will be converted to integers with no message. If wait times or claim frequencies
are less than zero, or payment frequencies are less than one, they will be converted with a message.
}
