% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolicyChars.R
\name{fill_addl}
\alias{fill_addl}
\title{Fill the additional characteristics of a simulated policy table, or create
a simulated policy table filling in additional characteristics from a pol_char object}
\usage{
fill_addl(plc, tbl_policy = NULL, ...)
}
\arguments{
\item{plc}{a \code{pol_char} object used to fill additional fields}

\item{tbl_policy}{A simulated policy table. If NULL then one will be created}

\item{...}{Used to for parameters to call \code{policies_simulate} if needed.
Do not include \code{additional_columns}; those will be filled with pol_char}
}
\description{
Fill the additional characteristics of a simulated policy table, or create
a simulated policy table filling in additional characteristics from a pol_char object
}
