% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherHelpers.R
\name{FixedHelper}
\alias{FixedHelper}
\title{FixedHelper}
\usage{
FixedHelper(Fixed)
}
\arguments{
\item{Fixed}{A scalar value}
}
\value{
A function
}
\description{
Returns the same, non-stochastic, value
}
