% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityHelpers.R
\name{GammaHelper}
\alias{GammaHelper}
\title{GammaHelper}
\usage{
GammaHelper(alpha, beta, SingletonList = FALSE)
}
\arguments{
\item{alpha}{Double for the \eqn{\alpha} parameter}

\item{beta}{Double for the \eqn{\beta} parameter}

\item{SingletonList}{Boolean indicating whether to return a single function as a list. Default FALSE.}
}
\description{
This will create a random number generator for the gamma function.
}
