% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{GrowPolicies}
\alias{GrowPolicies}
\title{Simulate policy growth}
\usage{
GrowPolicies(dfPolicy, Growth)
}
\arguments{
\item{dfPolicy}{Data frame of policy data}

\item{Growth}{Scalar value greater than or equal to zero}
}
\description{
Given a policy data frame, this will generate new policies in subsequent policy years.
}

