% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagine-package.r
\docType{package}
\name{imaginator}
\alias{imaginator}
\title{imaginator}
\description{
Simulate general insurance policies, losses and loss emergence. The package contemplates
deterministic and stochastic policy retention and growth scenarios. Retention and growth rates are percentages relative
to the expiring portfolio. Claims are simulated for each policy. This is accomplished either be assuming a frequency
distribution per development lag or by generating random wait times until claim emergence and settlement. Loss simulation
uses standard loss distributions for claim amounts.
}
