% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{IncrementPolicyYear}
\alias{IncrementPolicyYear}
\title{Incremental a policy year}
\usage{
IncrementPolicyYear(dfPolicy, Renewal, Growth)
}
\arguments{
\item{dfPolicy}{A policy data frame}

\item{Renewal}{Scalar renewal rate}

\item{Growth}{Scalar growth rate}
}
\value{
Policy data frame
}

