% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolicyChars.R
\name{is.pol_char}
\alias{is.pol_char}
\title{is.pol_char}
\usage{
is.pol_char(plc)
}
\arguments{
\item{plc}{A \code{pol_list} object}
}
\value{
Boolean value indicating if plc is a pol_char object
}
\description{
is.pol_char
}
