% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityHelper.R
\name{LognormalHelper}
\alias{LognormalHelper}
\title{LognormalHelper}
\usage{
LognormalHelper(meanlog, sdlog)
}
\arguments{
\item{meanlog}{Mean on the log scale}

\item{sdlog}{Standard deviation on the log scale}
}
\description{
This will create a random number generator for the lognormal function
}

