% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{NewPolicies}
\alias{NewPolicies}
\title{Simulate a new set of policies}
\usage{
NewPolicies(N, PolicyYear, Exposure = 1)
}
\arguments{
\item{N}{The number of policies to generate}

\item{PolicyYear}{Scalar integer indicating the policy year to generate}

\item{Exposure}{Vector of exposures}
}
\value{
Data frame of policy data
}
\details{
This function will create a data frame of policy data. Effective dates are uniformly distributed throughout the
year.
}

