% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{OneSimulation}
\alias{OneSimulation}
\title{OneSimulation}
\usage{
OneSimulation(AY = 0:5, freqFn = rnorm, freqArgs = list(40, 7.745967),
  occFn = runif, occArgs = list(1, 12), reportFn = rexp,
  reportArgs = list(1/18), payFn = rexp, payArgs = list(1/12),
  sevFn = rlnorm, sevArgs = list(8, 1.5815085))
}
\arguments{
\item{AY}{Vector of accident years}

\item{freq}{A function and arguments to simulate frequency}
}
\description{
Construct a data frame of simulated claims.
}

