% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyHelpers.R
\name{PoissonHelper}
\alias{PoissonHelper}
\title{PoissonHelper}
\usage{
PoissonHelper(lambda)
}
\arguments{
\item{lambda}{Expected value of poisson function}
}
\value{
A function
}
\description{
This will return a poisson random number generator with a user-defined mean.
}
