% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{policies_grow}
\alias{policies_grow}
\title{Simulate policy growth}
\usage{
policies_grow(tbl_policy, growth)
}
\arguments{
\item{tbl_policy}{Data frame of policy data}

\item{growth}{Scalar value greater than or equal to zero}
}
\value{
A data frame, as follows:
\describe{
  \item{policy_effective_date}{Date}
  \item{policy_expiration_date}{Date}
  \item{exposure}{double}
  \item{policyholder_id}{integer}
}
}
\description{
Given a policy data frame, this will generate new policies in subsequent policy years.
}
