% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{policy_year_increment}
\alias{policy_year_increment}
\title{Incremental a policy year}
\usage{
policy_year_increment(tbl_policy, retention, growth)
}
\arguments{
\item{tbl_policy}{A policy data frame}

\item{retention}{Scalar renewal rate}

\item{growth}{Scalar growth rate}
}
\value{
Policy data frame
}
\description{
Given a policy data frame, this will combine the \code{policies_grow} and \code{policies_renew} functions
to produce a subsequent policy year.
}
