% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{policy_year_new}
\alias{policy_year_new}
\title{Simulate a new policy year}
\usage{
policy_year_new(n, policy_year, exposure = 1, start_id = 1, additional_columns)
}
\arguments{
\item{n}{The number of policies to generate}

\item{policy_year}{Scalar integer indicating the policy year to generate}

\item{exposure}{Vector of exposures}

\item{start_id}{Integer of the first number in the policy ID sequence}

\item{additional_columns}{A list of addtional column names and values}
}
\value{
Data frame of policy data
}
\description{
This will generate a data frame of policy data. This may be used to construct renewal and growth
data frames for subsequent policy years.
}
\details{
Effective dates are uniformly distributed throughout the
year.

When providing additional columns, each element of the list must be a scalar and be named.
}
