% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateClaimsByEval.R
\name{ClaimsByEvaluationDate}
\alias{ClaimsByEvaluationDate}
\title{Claims by evaluation date}
\usage{
ClaimsByEvaluationDate(dfPolicy, Frequency, Severity, Links, EvaluationDates)
}
\arguments{
\item{dfPolicy}{A policy data frame}

\item{Frequency}{A function which will randomly generate number of claims}

\item{Severity}{A function which will randomly generate the severity of each claim}

\item{Links}{A list of functions which dictates how severities change from one evaluation date to the next}

\item{EvaluationDates}{A vector of evaluation dates}
}
\value{
A claims data frame
}
\description{
Given a data frame of policies, this will simulate claims
}
\details{
The claim amounts are presumed to be cumulative
}

