% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaimsByFirstReport.R
\name{ClaimsByFirstReport}
\alias{ClaimsByFirstReport}
\title{Claims by first report}
\usage{
ClaimsByFirstReport(dfPolicy, Frequency, PaymentSeverity, Lags)
}
\arguments{
\item{dfPolicy}{A policy data frame}

\item{Frequency}{A function, or list of functions which will randomly generate the number of claims per policy}

\item{PaymentSeverity}{A function, or list of functions which will randomly generate the payment for each claim}

\item{Lags}{A vector of lags as integers}
}
\value{
A claims data frame
}
\description{
Given a data frame of policies, this will simulate claims by applying a (possibly) random link ratio.
}
\details{
Creates a data frame with randomly generated claim values
}
