% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherHelpers.R
\name{FixedHelper}
\alias{FixedHelper}
\title{FixedHelper}
\usage{
FixedHelper(Fixed, SingletonList = FALSE)
}
\arguments{
\item{Fixed}{A scalar value}

\item{SingletonList}{Boolean indicating whether to return a single function as a list. Default FALSE.}
}
\value{
A function
}
\description{
Returns the same, non-stochastic, value
}
