% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityHelper.R
\name{GammaHelper}
\alias{GammaHelper}
\title{GammaHelper}
\usage{
GammaHelper(alpha, beta)
}
\arguments{
\item{alpha}{Double for the \eqn{\alpha} parameter}

\item{beta}{Double for the \eqn{\beta} parameter}
}
\description{
This will create a random number generator for the gamma function.
}
