% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{GrowPolicies}
\alias{GrowPolicies}
\alias{RenewPolicies}
\title{Simulate policy growth}
\usage{
RenewPolicies(dfPolicy, Renewal)
}
\arguments{
\item{dfPolicy}{Data frame of policy data}

\item{Renewal}{Scalar value between zero and one}
}

