% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{NewPolicies}
\alias{NewPolicies}
\title{Simulate a new set of policies}
\usage{
NewPolicies(N, PolicyYear, Exposure = 1, AdditionalColumns)
}
\arguments{
\item{N}{The number of policies to generate}

\item{PolicyYear}{Scalar integer indicating the policy year to generate}

\item{Exposure}{Vector of exposures}

\item{AdditionalColumns}{A list of addtional column names and values}
}
\value{
Data frame of policy data
}
\description{
This will generate a data frame of policy data. This may be used to construct renewal and growth
data frames for subsequent policy years.
}
\details{
Effective dates are uniformly distributed throughout the
year.

When providing additional columns, each element of the list must be a scalar and be named.
}

