% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherHelpers.R
\name{NormalHelper}
\alias{NormalHelper}
\title{NormalHelper}
\usage{
NormalHelper(mean, sd, lowerBound, upperBound)
}
\arguments{
\item{mean}{The mean}

\item{sd}{The standard deviation}

\item{lowerBound}{Lower boundary of the return values}

\item{upperBound}{Upper bound of the returned values}
}
\value{
A function
}
\description{
Returns a normal distribution with optional bounds on the returned values
}

