% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{SimulatePolicies}
\alias{SimulatePolicies}
\title{Simulate a data frame of policies}
\usage{
SimulatePolicies(StartYear, NumYears, N, Exposure = 1)
}
\arguments{
\item{N}{An integer giving the number of policies per year}

\item{Exposure}{Exposure per policy}

\item{PolicyYears}{A vector of integers in sequence}

\item{Growth}{A vector indicating the rate of growth of policies}

\item{Retention}{A vector indicating loss of policies}
}
\value{
A data frame of policy data as explained in
}
\details{

}
\examples{

py <- 2001:2010

N <- 5000

exposure <- 1
rate <- 1

retention <- 0.9
}

