% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePolicies.R
\name{SimulatePolicies}
\alias{SimulatePolicies}
\title{Simulate a data frame of policies}
\usage{
SimulatePolicies(N, PolicyYears, Exposure = 1, Renewal, Growth,
  AdditionalColumns)
}
\arguments{
\item{N}{An integer giving the number of policies in the first year}

\item{PolicyYears}{A vector of integers in sequence}

\item{Exposure}{Exposure per policy}

\item{Renewal}{A vector indicating loss of policies}

\item{Growth}{A vector indicating the rate of growth of policies}

\item{AdditionalColumns}{A list of addtional column names and values}
}
\value{
A data frame of policy data
}
\description{
Given a starting number of policies, this function will generate additional years of policy data.
}

