% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OtherHelpers.R
\name{UniformHelper}
\alias{UniformHelper}
\title{UniformHelper}
\usage{
UniformHelper(min, max)
}
\arguments{
\item{min}{The minimum value}

\item{max}{The maximum value}
}
\value{
A function
}
\description{
Returns a uniform distribution function
}

